Restructure Merkle verification types
